minimum viability
