--- src/LlmTornado.A2A.AgentServer/A2ATornadoRuntimeConfiguration.cs
@@ -1,290 +1,0 @@
-﻿using A2A;
-using A2A.AspNetCore;
-using LlmTornado.Agents;
-using LlmTornado.Agents.ChatRuntime;
-using LlmTornado.Agents.ChatRuntime.RuntimeConfigurations;
-using LlmTornado.Agents.DataModels;
-using LlmTornado.Chat;
-using Microsoft.AspNetCore.Mvc.ApplicationModels;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using System.Collections.Concurrent;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Text.Json;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace LlmTornado.A2A.AgentServer;
-
-public interface IA2ARuntimeConfiguration
-{
-    Task<A2AResponse> ProcessMessageAsync(MessageSendParams messageSendParams, CancellationToken cancellationToken);
-    Task ExecuteAgentTaskAsync(AgentTask task, CancellationToken cancellationToken);
-    Task<AgentCard> GetAgentCardAsync(string agentUrl, CancellationToken cancellationToken);
-}
-
-/// <summary>
-/// Wraps Semantic Kernel-based agents to handle Travel related tasks
-/// </summary>
-public class A2ATornadoRuntimeConfiguration : IA2ARuntimeConfiguration, IDisposable
-{
-    public static readonly ActivitySource ActivitySource = new("A2A.TornadoRuntimeAgent", "1.0.0");
-
-    //private readonly ILogger _logger;
-    private readonly ChatRuntime _agent;
-    private ITaskManager? _taskManager;
-    private IRuntimeConfiguration _runtimeConfig;
-    private AgentTask _currentTask;
-    private ConcurrentQueue<Artifact> _artifactQueue = new();
-    private CancellationToken _cancellationToken;
-
-    /// <summary>
-    /// Initializes a new instance of the A2ATornadoRuntimeService
-    /// </summary>
-    public A2ATornadoRuntimeConfiguration(IRuntimeConfiguration runtimeConfig)
-    {
-        //_logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        _runtimeConfig = runtimeConfig ?? throw new ArgumentNullException(nameof(runtimeConfig));
-
-        // Initialize the agent
-        _agent = new ChatRuntime(_runtimeConfig);
-
-        //Setup streaming event handler
-        _agent.RuntimeConfiguration.OnRuntimeEvent += ProcessRuntimeStreamingEvent;
-        _agent.RuntimeConfiguration.OnRuntimeEvent += ProcessRuntimeOrchestrationEvent;
-    }
-
-    /// <summary>
-    /// Dispose of resources
-    /// </summary>
-    public void Dispose()
-    {
-        GC.SuppressFinalize(this);
-    }
-
-    /// <summary>
-    /// Interface to the TaskManager
-    /// </summary>
-    /// <param name="taskManager"></param>
-    public void Attach(ITaskManager taskManager)
-    {
-        _taskManager = taskManager;
-        taskManager.OnTaskCreated = ExecuteAgentTaskAsync;
-        taskManager.OnTaskUpdated = ExecuteAgentTaskAsync;
-        taskManager.OnAgentCardQuery = GetAgentCardAsync;
-    }
-
-    /// <summary>
-    /// Process a message sent to the agent Without streaming
-    /// </summary>
-    /// <param name="messageSendParams"></param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    public async Task<A2AResponse> ProcessMessageAsync(MessageSendParams messageSendParams, CancellationToken cancellationToken)
-    {
-        //Check for cancellation
-        if (cancellationToken.IsCancellationRequested)
-        {
-            return new AgentMessage()
-            {
-                Role = MessageRole.Agent,
-                MessageId = Guid.NewGuid().ToString(),
-                ContextId = messageSendParams.Message.ContextId,
-                Parts = [new TextPart() {
-                    Text = "Operation cancelled."
-                }]
-            };
-        }
-
-        // Get the message
-        var messageText = messageSendParams.Message.Parts.OfType<TextPart>().First().Text;
-
-        //Invoke Runtime
-        ChatMessage response = await _agent.InvokeAsync(new ChatMessage(Code.ChatMessageRoles.User, messageText));
-
-        // Create and return an artifact
-        return response.ToA2AAgentMessage();
-    }
-
-    /// <summary>
-    /// Process a task assigned to the agent with streaming output
-    /// </summary>
-    /// <param name="task"></param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    /// <exception cref="InvalidOperationException"></exception>
-    public async Task ExecuteAgentTaskAsync(AgentTask task, CancellationToken cancellationToken)
-    {
-        if (_taskManager == null)
-        {
-            throw new InvalidOperationException("TaskManager is not attached.");
-        }
-
-        //Shared for the event handler
-        _cancellationToken = cancellationToken;
-
-        //Set current task
-        _currentTask = task;
-
-        //Check for cancellation
-        if (cancellationToken.IsCancellationRequested)
-        {
-            await _taskManager.UpdateStatusAsync(task.Id, TaskState.Canceled, 
-                message: new AgentMessage()
-                {
-                    Role = MessageRole.Agent,
-                    MessageId = Guid.NewGuid().ToString(),
-                    Parts = [new TextPart() {
-                        Text = "Operation cancelled."
-                    }]
-                },
-            cancellationToken: cancellationToken);
-            return;
-        }
-
-        //Send Notify working status
-        await _taskManager.UpdateStatusAsync(task.Id, TaskState.Submitted, cancellationToken: cancellationToken);
-        await _taskManager.UpdateStatusAsync(task.Id, TaskState.Working, cancellationToken: cancellationToken);
-
-        // Get message from the user
-        var userMessage = task.History!.Last().Parts.OfType<TextPart>().First().Text;
-
-
-        // Get the response from the agent
-        ChatMessage response = await _agent.InvokeAsync(new ChatMessage(Code.ChatMessageRoles.User, userMessage ?? "Empty message"));
-
-        var artifact = new Artifact()
-        {
-            ArtifactId = response.Id.ToString(),
-            Description = "response from agent",
-            Parts = [new TextPart() {Text = response.Content ?? "" }]
-        };
-
-        await _taskManager.ReturnArtifactAsync(_currentTask.Id, artifact, _cancellationToken);
-
-        await _taskManager.UpdateStatusAsync(task.Id, TaskState.Working, message: response.ToA2AAgentMessage(), cancellationToken: cancellationToken);
-
-        // Update the Status to Completed
-        await _taskManager.UpdateStatusAsync(_currentTask.Id, TaskState.Completed, final:true, cancellationToken: cancellationToken);
-    }
-
-    /// <summary>
-    /// Process runtime events for streaming output
-    /// </summary>
-    /// <param name="evt"></param>
-    /// <returns></returns>
-    private async ValueTask ProcessRuntimeStreamingEvent(ChatRuntimeEvents evt)
-    {
-        if (evt.EventType == ChatRuntimeEventTypes.AgentRunner)
-        {
-            if (evt is ChatRuntimeAgentRunnerEvents runnerEvt)
-            {
-                if (runnerEvt.AgentRunnerEvent is AgentRunnerStreamingEvent streamEvt)
-                {
-                    if (streamEvt.ModelStreamingEvent is ModelStreamingOutputTextDeltaEvent deltaTextEvent)
-                    {
-                        Artifact artifact = new Artifact()
-                        {
-                            ArtifactId = Guid.NewGuid().ToString(),
-                            Description = "Streaming event",
-                            Parts = [new TextPart() {
-                                Text = deltaTextEvent.DeltaText ?? ""
-                            }]
-                        };
-                        Console.Write(deltaTextEvent.DeltaText);
-                        await _taskManager.ReturnArtifactAsync(_currentTask.Id, artifact, _cancellationToken);
-                        await Task.Delay(100); // Adjust the delay as needed
-                    }
-                }
-            }
-        }
-    }
-
-
-
-    /// <summary>
-    /// Possibly handle logging
-    /// </summary>
-    /// <param name="evt"></param>
-    /// <returns></returns>
-    private async ValueTask ProcessRuntimeOrchestrationEvent(ChatRuntimeEvents evt)
-    {
-        if (evt.EventType == ChatRuntimeEventTypes.Orchestration)
-        {
-            if (evt is ChatRuntimeOrchestrationEvent orchestrationEvt)
-            {
-                if (orchestrationEvt.OrchestrationEventData is OnVerboseOrchestrationEvent verbose)
-                {
-                    Artifact artifact = new Artifact()
-                    {
-                        ArtifactId = Guid.NewGuid().ToString(),
-                        Description = "Agent Orchestration Event",
-                        Parts = [new TextPart() {
-                                Text = verbose.Message ?? ""
-                            }]
-                    };
-                    await _taskManager.ReturnArtifactAsync(_currentTask.Id, artifact, _cancellationToken);
-                    await Task.Delay(100); // Adjust the delay as needed
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// Generic Agent Card for the agent
-    /// </summary>
-    /// <param name="agentUrl"></param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    public Task<AgentCard> GetAgentCardAsync(string agentUrl, CancellationToken cancellationToken)
-    {
-        if (cancellationToken.IsCancellationRequested)
-        {
-            return Task.FromCanceled<AgentCard>(cancellationToken);
-        }
-
-        return Task.FromResult(DescribeAgentCard(agentUrl));
-    }
-
-
-    /// <summary>
-    /// Defines a static Agent Card for the agent
-    /// </summary>
-    /// <returns></returns>
-    public virtual AgentCard DescribeAgentCard(string agentUrl)
-    {
-        AgentCapabilities capabilities = new AgentCapabilities()
-        {
-            Streaming = false,
-            PushNotifications = false,
-        };
-
-        AgentSkill chattingSkill = new AgentSkill()
-        {
-            Id = "chatting_skill",
-            Name = "Chatting feature",
-            Description = "Agent to chat with and search the web.",
-            Tags = ["chat", "websearch", "llm-tornado"],
-            Examples =
-            [
-                "Hello, what's up?",
-                "What is the weather like in boston?",
-            ],
-        };
-
-        return new AgentCard()
-        {
-            Name = "Tornado Agent",
-            Description = "Agent to chat with and search the web",
-            Url = agentUrl, // Placeholder URL
-            Version = "1.0.0",
-            DefaultInputModes = ["text"],
-            DefaultOutputModes = ["text"],
-            Capabilities = capabilities,
-            Skills = [chattingSkill],
-        };
-    }   
-}
-
-
