+++ src/LlmTornado.A2A.AgentServer/Agent/ChatbotAgent.cs
@@ -6,18 +6,8 @@ using LlmTornado.Agents.DataModels;
 using LlmTornado.Chat;
 using LlmTornado.Chat.Models;
 using LlmTornado.Code;
-using LlmTornado.Embedding;
-using LlmTornado.Embedding.Models;
-using LlmTornado.Infra;
 using LlmTornado.Moderation;
 using LlmTornado.Responses;
-
-using Newtonsoft.Json;
-using System.Collections.Concurrent;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
-using System.Text.Json.Serialization;
 
 namespace LlmTornado.A2A.AgentServer;
