+++ src/LlmTornado.A2A.AgentServer/ReadMe.md
@@ -1,0 +1,7 @@
+﻿# Using The A2A Agent Server
+
+1) Create your agent
+2) Add your agent on `Program.cs`
+3) Build the project
+4) Build the docker image
+5) Run the docker image `docker run -it -p 5000:80 your-image-name`
