+++ src/LlmTornado.A2A.Hosting/Services/A2AContainerService.cs
@@ -3,15 +3,6 @@ using System.Net.ServerSentEvents;
 
 namespace LlmTornado.A2A.Hosting.Services;
 
-public interface IA2AContainerService
-{
-    Task<AgentCard> GetAgentCardAsync(string endPoint);
-    Task<A2AResponse> SendMessageAsync(string endPoint, List<Part> parts);
-    Task SendMessageStreamingAsync(string endPoint, List<Part> parts, Func<SseItem<A2AEvent>, Task>? onEventReceived);
-    Task<AgentTask> CancelTaskAsync(string endPoint, string taskId);
-    Task<AgentTask> GetTaskAsync(string endPoint, string taskId);
-}
-
 public class A2AContainerService : IA2AContainerService
 {
     public async Task<AgentCard> GetAgentCardAsync(string endPoint)
