+++ src/LlmTornado.A2A/LlmTornado.A2A.csproj
@@ -23,6 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\LlmTornado.Agents\LlmTornado.Agents.csproj" />
     <ProjectReference Include="..\LlmTornado\LlmTornado.csproj" />
   </ItemGroup>
